requirements.txt
alembic==1.8.1
bcrypt==4.0.1
blinker==1.5
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==1.1.3.post0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
Pillow==9.2.0
PyJWT==2.6.0
PyMySQL==1.0.2
SQLAlchemy==1.4.42
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.9.0
